+++ tiny_dnn/core/kernels/maxpool_op_internal.h
@@ -15,7 +15,7 @@ inline void maxpool_op_internal(const tensor_t &in_data,
                                 std::vector<std::vector<size_t>> &max_idx,
                                 const std::vector<std::vector<size_t>> &out2in,
                                 const bool layer_parallelize) {
-  for_i(layer_parallelize, in_data.size(), [&](int sample) {
+  for_i(layer_parallelize, in_data.size(), [&](size_t sample) {
     const vec_t &in          = in_data[sample];
     vec_t &out               = out_data[sample];
     std::vector<size_t> &max = max_idx[sample];
@@ -41,7 +41,7 @@ inline void maxpool_grad_op_internal(tensor_t &prev_delta,
                                      std::vector<std::vector<size_t>> &max_idx,
                                      const std::vector<size_t> &in2out,
                                      const bool layer_parallelize) {
-  for_i(layer_parallelize, prev_delta.size(), [&](int sample) {
+  for_i(layer_parallelize, prev_delta.size(), [&](size_t sample) {
     vec_t &prev                    = prev_delta[sample];
     const vec_t &curr              = curr_delta[sample];
     const std::vector<size_t> &max = max_idx[sample];
